Separate state var generation part as a new class, TSStateGenerator.
